// src/interfaces/InterfaceMovieDetails.ts
export interface InterfaceMovieDetails {
	adult: boolean;
	backdrop_path: string;
	belongs_to_collection: Belongstocollection;
	budget: number;
	genres: Genre[];
	homepage: string;
	id: number;
	imdb_id: string;
	origin_country: string[];
	original_language: string;
	original_title: string;
	overview: string;
	popularity: number;
	poster_path: string;
	production_companies?: Productioncompanies[];
	production_countries: Productioncountry[];
	release_date: string;
	revenue: number;
	runtime: number;
	spoken_languages: Spokenlanguage[];
	status: string;
	tagline: string;
	title: string;
	video: boolean;
	vote_average: number;
	vote_count: number;
}

interface Spokenlanguage {
	english_name: string;
	iso_639_1: string;
	name: string;
}

interface Productioncountry {
	iso_3166_1: string;
	name: string;
}

interface Genre {
	id: number;
	name: string;
}

interface Productioncompanies {
	id: number;
	logo_path: string;
	name: string;
	origin_country: string;
}

interface Belongstocollection {
	id: number;
	name: string;
	poster_path: string;
	backdrop_path: string;
}